Rename test.py to contract_utils.py
Btw. Add docstring to the call_validation_code
